+++ src/com/walking/lesson1_varaibles/Task1.java
@@ -1,4 +1,4 @@
-package com.walking.lesson1;
+package com.walking.lesson1_varaibles;
 
 /**
  * На месте комментария написать код, который присвоит переменной a значение переменной b и наоборот.
