+++ src/com/walking/lesson3_casts_conditional_constructions/Task2SwitchCase.java
@@ -1,4 +1,4 @@
-package com.walking.lesson3;
+package com.walking.lesson3_casts_conditional_constructions;
 
 /**
  * Написать программу, которая принимает строку с клавиатуры.
