+++ src/com/walking/lesson3_casts_conditional_constructions/Task3Printf.java
@@ -1,12 +1,12 @@
-package com.walking.lesson3;
+package com.walking.lesson3_casts_conditional_constructions;
 
 /**
  * Завести три переменные типа String.
  * С клавиатуры ввести в них вашу фамилию, имя и отчество.
  * Вывести в консоль в одну строку ваше ФИО.
- * Сделать, используя System.out.println
+ * Сделать, используя System.out.printf
  */
-public class Task3Println {
+public class Task3Printf {
     public static void main(String[] args) {
     }
 }
