+++ src/com/walking/lesson3_casts_conditional_constructions/Task3Println.java
@@ -1,12 +1,12 @@
-package com.walking.lesson3;
+package com.walking.lesson3_casts_conditional_constructions;
 
 /**
  * Завести три переменные типа String.
  * С клавиатуры ввести в них вашу фамилию, имя и отчество.
  * Вывести в консоль в одну строку ваше ФИО.
- * Сделать, используя System.out.printf
+ * Сделать, используя System.out.println
  */
-public class Task3Printf {
+public class Task3Println {
     public static void main(String[] args) {
     }
 }
