+++ src/com/walking/lesson3_casts_conditional_constructions/Task4.java
@@ -1,4 +1,4 @@
-package com.walking.lesson3;
+package com.walking.lesson3_casts_conditional_constructions;
 
 /**
  * Ввести с клавиатуры два целых числа.
